base: Use a system-wide recents corner radius
With Google implementation of quickstep, the rounded corners are set indivdually according to the view.
If it's the AOSP default recents, it'll take a value for rounded corners.
If it's Launcher3 or any other app, it'll take a custom value that's declared on the app.

Let's make a public value and use it on both because our themes need consistency