policy: Add support for tri-state key events
This adds support for alert sliders that base their states on key codes,
specifically newer oneplus devices. We've previously read the state path
on Oneplus 5 and below but Oneplus 6 and beyond seems to trigger their
states via keycodes. This just adds a secondary layer to support the
newer devices as well.

This requires config_hasAlertSlider and all 3 keycode configs to be
greater than 0 (matching the devices keycodes). This will not interfere
with our older state path implementation since that checks for path
overlays as well